add return type check for glob

+++ src/SessionStorage.php
@@ -144,7 +144,11 @@ class SessionStorage implements SessionStorageInterface
         $expiresAt = clone $this->dateTime;
         $expiresAt->sub($expiresIn);
         $sessionFileFilter = \sprintf('%s/%s*', $this->sessionDir, self::SESSION_FILE_PREFIX);
-        foreach (\glob($sessionFileFilter) as $sessionFile) {
+        if (false === $sessionFileList = \glob($sessionFileFilter)) {
+            // unable to list files in sessionDir, nothing we can do...
+            return;
+        }
+        foreach ($sessionFileList as $sessionFile) {
             $lastModified = \filemtime($sessionFile);
             if ($lastModified < $expiresAt->getTimestamp()) {
                 \unlink($sessionFile);
